# docs/openapi/claims-api.yaml
openapi: 3.0.3
info:
  title: Claims Processing API
  version: 1.0.0
  description: Executable contract for the local claims-processing prototype.
servers:
  - url: http://localhost:8080
paths:
...
  /internal/v1/payment-status-events:
    post:
      operationId: receivePaymentStatus
      parameters:
        - name: X-Callback-Timestamp
          in: header
          required: true
          schema:
            type: string
            description: Unix epoch timestamp in seconds.
        - name: X-Callback-Signature
          in: header
          required: true
          schema:
            type: string
            pattern: '^v1=[0-9a-f]{64}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentStatusEvent'
      responses:
        '202':
          description: Event authenticated and accepted idempotently.
        '401':
          $ref: '#/components/responses/Problem'
components:
...
  responses:
    Problem:
      description: RFC 9457 problem detail.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    PaymentStatusEvent:
      type: object
      required: [eventId, claimId, paymentReference, status, occurredAt]
      properties:
        eventId: {type: string, format: uuid}
        claimId: {type: string, format: uuid}
        paymentReference: {type: string, minLength: 1}
        status: {type: string, enum: [COMPLETED, FAILED]}
        occurredAt: {type: string, format: date-time}
    Problem:
      type: object
      properties:
        status: {type: integer}
        title: {type: string}
        detail: {type: string}
        code: {type: string}